requirements.txt
absl-py==0.10.0
appdirs==1.4.4
applicationinsights==0.11.9
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.9.1
bleach==3.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cli-helpers==0.2.3
click==7.1.2
configobj==5.0.6
convertapi==1.4.0
crcmod==1.7
cryptography==3.1.1
decorator==4.4.2
distlib==0.3.1
gast==0.4.0
googletrans==3.0.0
gTTS==2.1.1
gTTS-token==1.1.3
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.10.6
html-text==0.5.2
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
humanize==3.1.0
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==2.0.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
libcst==0.3.13
lxml==4.6.1
Markdown==3.3.2
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==1.0.3
mutagen==1.45.1
mypy==0.790
mypy-extensions==0.4.3
numpy==1.19.2
oauth2client==4.1.3
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.17.3
pytesseract==0.3.6
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
pyzbar==0.1.8
requests==2.24.0
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
snowballstemmer==2.0.0
soupsieve==2.0.1
speedtest-cli==2.1.2
sqlparse==0.2.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
terminaltables==3.1.0
toml==0.10.1
tqdm==4.50.2
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
yapf==0.30.0
zipp==3.3.2
openwa
